[Infra UI] Avoid eager async imports in metric alert registrations (#123285)
(cherry picked from commit 629c602a67e090c1c5a7e7cbb61206d511f0845a)

# Conflicts:
#	x-pack/plugins/infra/public/alerting/inventory/components/expression.tsx
#	x-pack/plugins/infra/public/alerting/inventory/components/metric.tsx
#	x-pack/plugins/infra/public/alerting/metric_anomaly/components/expression.tsx
#	x-pack/plugins/infra/public/alerting/metric_threshold/components/expression.tsx
#	x-pack/plugins/infra/public/alerting/metric_threshold/components/expression_row.tsx
#	x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/evaluate_condition.ts
#	x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/lib/calculate_from_based_on_metric.ts
#	x-pack/plugins/infra/server/lib/alerting/metric_threshold/lib/evaluate_rule.ts
#	x-pack/plugins/infra/server/lib/alerting/metric_threshold/register_metric_threshold_rule_type.ts
#	x-pack/test/api_integration/apis/metrics_ui/inventory_threshold_alert.ts
