+++ x-pack/plugins/infra/public/alerting/inventory/components/metric.tsx
@@ -5,34 +5,34 @@
  * 2.0.
  */
 
-import React, { useState, useCallback, useMemo } from 'react';
-import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n-react';
-import { debounce } from 'lodash';
 import {
+  EuiButtonGroup,
+  EuiButtonIcon,
+  EuiComboBox,
   EuiExpression,
-  EuiPopover,
+  EuiFieldText,
   EuiFlexGroup,
   EuiFlexItem,
   EuiFormRow,
-  EuiComboBox,
-  EuiButtonGroup,
-  EuiSpacer,
+  EuiPopover,
+  EuiPopoverTitle,
   EuiSelect,
+  EuiSpacer,
   EuiText,
-  EuiFieldText,
 } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
+import { FormattedMessage } from '@kbn/i18n-react';
+import { debounce } from 'lodash';
+import React, { useCallback, useMemo, useState } from 'react';
 import { IFieldType } from 'src/plugins/data/public';
-import { EuiPopoverTitle, EuiButtonIcon } from '@elastic/eui';
+import { IErrorObject } from '../../../../../triggers_actions_ui/public';
 import { getCustomMetricLabel } from '../../../../common/formatters/get_custom_metric_label';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { IErrorObject } from '../../../../../triggers_actions_ui/public/types';
 import {
+  SnapshotCustomAggregation,
+  SnapshotCustomAggregationRT,
   SnapshotCustomMetricInput,
   SnapshotCustomMetricInputRT,
-  SnapshotCustomAggregation,
   SNAPSHOT_CUSTOM_AGGREGATIONS,
-  SnapshotCustomAggregationRT,
 } from '../../../../common/http_api/snapshot_api';
 
 interface Props {
