+++ x-pack/plugins/infra/public/alerting/inventory/components/validation.tsx
@@ -6,21 +6,20 @@
  */
 
 import { i18n } from '@kbn/i18n';
+import type { ValidationResult } from '../../../../../triggers_actions_ui/public';
 import {
-  InventoryMetricConditions,
   Comparator,
-  // eslint-disable-next-line @kbn/eslint/no-restricted-paths
-} from '../../../../server/lib/alerting/inventory_metric_threshold/types';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { ValidationResult } from '../../../../../triggers_actions_ui/public/types';
-import { QUERY_INVALID } from './expression';
+  FilterQuery,
+  InventoryMetricConditions,
+  QUERY_INVALID,
+} from '../../../../common/alerting/metrics';
 
 export function validateMetricThreshold({
   criteria,
   filterQuery,
 }: {
   criteria: InventoryMetricConditions[];
-  filterQuery?: string | symbol;
+  filterQuery?: FilterQuery;
 }): ValidationResult {
   const validationResult = { errors: {} };
   const errors: {
