+++ x-pack/plugins/infra/public/alerting/metric_anomaly/components/validation.tsx
@@ -6,8 +6,7 @@
  */
 
 import { i18n } from '@kbn/i18n';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { ValidationResult } from '../../../../../triggers_actions_ui/public/types';
+import type { ValidationResult } from '../../../../../triggers_actions_ui/public';
 
 export function validateMetricAnomaly({
   hasInfraMLCapabilities,
