+++ x-pack/plugins/infra/public/alerting/metric_threshold/components/expression.test.tsx
@@ -6,14 +6,13 @@
  */
 
 import { mountWithIntl, nextTick } from '@kbn/test/jest';
+import React from 'react';
+import { act } from 'react-dom/test-utils';
 // We are using this inside a `jest.mock` call. Jest requires dynamic dependencies to be prefixed with `mock`
 import { coreMock as mockCoreMock } from 'src/core/public/mocks';
+import { Comparator } from '../../../../common/alerting/metrics';
 import { MetricsExplorerMetric } from '../../../../common/http_api/metrics_explorer';
-import React from 'react';
 import { Expressions } from './expression';
-import { act } from 'react-dom/test-utils';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { Comparator } from '../../../../server/lib/alerting/metric_threshold/types';
 
 jest.mock('../../../containers/metrics_source/use_source_via_http', () => ({
   useSourceViaHttp: () => ({
