+++ x-pack/plugins/infra/public/alerting/metric_threshold/components/expression_chart.test.tsx
@@ -5,17 +5,16 @@
  * 2.0.
  */
 
+import { DataViewBase } from '@kbn/es-query';
 import { mountWithIntl, nextTick } from '@kbn/test/jest';
+import React from 'react';
+import { act } from 'react-dom/test-utils';
 // We are using this inside a `jest.mock` call. Jest requires dynamic dependencies to be prefixed with `mock`
 import { coreMock as mockCoreMock } from 'src/core/public/mocks';
-import { MetricExpression } from '../types';
-import { DataViewBase } from '@kbn/es-query';
+import { Aggregators, Comparator } from '../../../../common/alerting/metrics';
 import { MetricsSourceConfiguration } from '../../../../common/metrics_sources';
-import React from 'react';
+import { MetricExpression } from '../types';
 import { ExpressionChart } from './expression_chart';
-import { act } from 'react-dom/test-utils';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { Aggregators, Comparator } from '../../../../server/lib/alerting/metric_threshold/types';
 
 const mockStartServices = mockCoreMock.createStart();
 jest.mock('../../../hooks/use_kibana', () => ({
