+++ x-pack/plugins/infra/public/alerting/metric_threshold/types.ts
@@ -5,12 +5,9 @@
  * 2.0.
  */
 
-import {
-  MetricExpressionParams,
-  // eslint-disable-next-line @kbn/eslint/no-restricted-paths
-} from '../../../server/lib/alerting/metric_threshold/types';
-import { MetricsExplorerOptions } from '../../pages/metrics/metrics_explorer/hooks/use_metrics_explorer_options';
+import { FilterQuery, MetricExpressionParams } from '../../../common/alerting/metrics';
 import { MetricsExplorerSeries } from '../../../common/http_api/metrics_explorer';
+import { MetricsExplorerOptions } from '../../pages/metrics/metrics_explorer/hooks/use_metrics_explorer_options';
 
 export interface AlertContextMeta {
   currentOptions?: Partial<MetricsExplorerOptions>;
@@ -57,7 +54,7 @@ export interface ExpressionChartData {
 export interface AlertParams {
   criteria: MetricExpression[];
   groupBy?: string | string[];
-  filterQuery?: string | symbol;
+  filterQuery?: FilterQuery;
   sourceId: string;
   filterQueryText?: string;
   alertOnNoData?: boolean;
