+++ x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/evaluate_condition.ts
@@ -5,25 +5,26 @@
  * 2.0.
  */
 
-import moment from 'moment';
 import { ElasticsearchClient } from 'kibana/server';
-import { mapValues, last, first } from 'lodash';
+import { first, last, mapValues } from 'lodash';
+import moment from 'moment';
 import {
+  Comparator,
+  InventoryMetricConditions,
   isTooManyBucketsPreviewException,
   TOO_MANY_BUCKETS_PREVIEW_EXCEPTION,
 } from '../../../../common/alerting/metrics';
-import { InfraDatabaseSearchResponse, CallWithRequestParams } from '../../adapters/framework';
-import { Comparator, InventoryMetricConditions } from './types';
-import { InventoryItemType, SnapshotMetricType } from '../../../../common/inventory_models/types';
 import {
   InfraTimerangeInput,
-  SnapshotRequest,
   SnapshotCustomMetricInput,
+  SnapshotRequest,
 } from '../../../../common/http_api';
-import { InfraSource } from '../../sources';
-import { UNGROUPED_FACTORY_KEY } from '../common/utils';
+import { InventoryItemType, SnapshotMetricType } from '../../../../common/inventory_models/types';
 import { getNodes } from '../../../routes/snapshot/lib/get_nodes';
 import { LogQueryFields } from '../../../services/log_queries/get_log_query_fields';
+import { CallWithRequestParams, InfraDatabaseSearchResponse } from '../../adapters/framework';
+import { InfraSource } from '../../sources';
+import { UNGROUPED_FACTORY_KEY } from '../common/utils';
 
 type ConditionResult = InventoryMetricConditions & {
   shouldFire: boolean[];
