+++ x-pack/plugins/infra/server/lib/alerting/inventory_metric_threshold/register_inventory_metric_threshold_rule_type.ts
@@ -5,37 +5,40 @@
  * 2.0.
  */
 
-import { schema, Type } from '@kbn/config-schema';
 import { Unit } from '@elastic/datemath';
+import { schema, Type } from '@kbn/config-schema';
 import { i18n } from '@kbn/i18n';
 import { PluginSetupContract } from '../../../../../alerting/server';
+import {
+  Comparator,
+  METRIC_INVENTORY_THRESHOLD_ALERT_TYPE_ID,
+} from '../../../../common/alerting/metrics';
+import {
+  SnapshotCustomAggregation,
+  SNAPSHOT_CUSTOM_AGGREGATIONS,
+} from '../../../../common/http_api/snapshot_api';
+import {
+  InventoryItemType,
+  SnapshotMetricType,
+  SnapshotMetricTypeKeys,
+} from '../../../../common/inventory_models/types';
+import { InfraBackendLibs } from '../../infra_types';
+import {
+  alertStateActionVariableDescription,
+  groupActionVariableDescription,
+  metricActionVariableDescription,
+  reasonActionVariableDescription,
+  thresholdActionVariableDescription,
+  timestampActionVariableDescription,
+  valueActionVariableDescription,
+} from '../common/messages';
+import { oneOfLiterals, validateIsStringElasticsearchJSONFilter } from '../common/utils';
 import {
   createInventoryMetricThresholdExecutor,
   FIRED_ACTIONS,
   FIRED_ACTIONS_ID,
   WARNING_ACTIONS,
 } from './inventory_metric_threshold_executor';
-import { METRIC_INVENTORY_THRESHOLD_ALERT_TYPE_ID, Comparator } from './types';
-import { InfraBackendLibs } from '../../infra_types';
-import { oneOfLiterals, validateIsStringElasticsearchJSONFilter } from '../common/utils';
-import {
-  groupActionVariableDescription,
-  alertStateActionVariableDescription,
-  reasonActionVariableDescription,
-  timestampActionVariableDescription,
-  valueActionVariableDescription,
-  metricActionVariableDescription,
-  thresholdActionVariableDescription,
-} from '../common/messages';
-import {
-  SnapshotMetricTypeKeys,
-  SnapshotMetricType,
-  InventoryItemType,
-} from '../../../../common/inventory_models/types';
-import {
-  SNAPSHOT_CUSTOM_AGGREGATIONS,
-  SnapshotCustomAggregation,
-} from '../../../../common/http_api/snapshot_api';
 
 const condition = schema.object({
   threshold: schema.arrayOf(schema.number()),
