+++ x-pack/plugins/infra/server/lib/alerting/metric_threshold/metric_threshold_executor.test.ts
@@ -5,30 +5,30 @@
  * 2.0.
  */
 
-import { createMetricThresholdExecutor, FIRED_ACTIONS } from './metric_threshold_executor';
-import * as mocks from './test_mocks';
-// import { RecoveredActionGroup } from '../../../../../alerting/common';
-import {
-  alertsMock,
-  AlertServicesMock,
-  AlertInstanceMock,
-} from '../../../../../alerting/server/mocks';
-import { LifecycleAlertServices } from '../../../../../rule_registry/server';
-import { ruleRegistryMocks } from '../../../../../rule_registry/server/mocks';
-import { createLifecycleRuleExecutorMock } from '../../../../../rule_registry/server/utils/create_lifecycle_rule_executor_mock';
-import { InfraSources } from '../../sources';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { elasticsearchClientMock } from 'src/core/server/elasticsearch/client/mocks';
 import {
   AlertInstanceContext as AlertContext,
   AlertInstanceState as AlertState,
 } from '../../../../../alerting/server';
+// import { RecoveredActionGroup } from '../../../../../alerting/common';
+import {
+  AlertInstanceMock,
+  AlertServicesMock,
+  alertsMock,
+} from '../../../../../alerting/server/mocks';
+import { LifecycleAlertServices } from '../../../../../rule_registry/server';
+import { ruleRegistryMocks } from '../../../../../rule_registry/server/mocks';
+import { createLifecycleRuleExecutorMock } from '../../../../../rule_registry/server/utils/create_lifecycle_rule_executor_mock';
 import {
   Aggregators,
   Comparator,
   CountMetricExpressionParams,
   NonCountMetricExpressionParams,
-} from './types';
+} from '../../../../common/alerting/metrics';
+import { InfraSources } from '../../sources';
+import { createMetricThresholdExecutor, FIRED_ACTIONS } from './metric_threshold_executor';
+import * as mocks from './test_mocks';
 
 interface AlertTestInstance {
   instance: AlertInstanceMock;
