+++ x-pack/test/api_integration/apis/metrics_ui/inventory_threshold_alert.ts
@@ -10,12 +10,12 @@ import { convertToKibanaClient } from '@kbn/test';
 import {
   Comparator,
   InventoryMetricConditions,
-} from '../../../../plugins/infra/server/lib/alerting/inventory_metric_threshold/types';
+} from '../../../../plugins/infra/common/alerting/metrics';
+import { InventoryItemType } from '../../../../plugins/infra/common/inventory_models/types';
+import { evaluateCondition } from '../../../../plugins/infra/server/lib/alerting/inventory_metric_threshold/evaluate_condition';
 import { InfraSource } from '../../../../plugins/infra/server/lib/sources';
 import { FtrProviderContext } from '../../ftr_provider_context';
 import { DATES } from './constants';
-import { evaluateCondition } from '../../../../plugins/infra/server/lib/alerting/inventory_metric_threshold/evaluate_condition';
-import { InventoryItemType } from '../../../../plugins/infra/common/inventory_models/types';
 
 export default function ({ getService }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
