+++ x-pack/test/api_integration/apis/metrics_ui/metric_threshold_alert.ts
@@ -7,18 +7,18 @@
 
 import expect from '@kbn/expect';
 import { convertToKibanaClient } from '@kbn/test';
-import { InfraSource } from '../../../../plugins/infra/common/source_configuration/source_configuration';
-import { FtrProviderContext } from '../../ftr_provider_context';
-import {
-  evaluateRule,
-  EvaluatedRuleParams,
-} from '../../../../plugins/infra/server/lib/alerting/metric_threshold/lib/evaluate_rule';
 import {
   Aggregators,
+  Comparator,
   CountMetricExpressionParams,
   NonCountMetricExpressionParams,
-} from '../../../../plugins/infra/server/lib/alerting/metric_threshold/types';
-import { Comparator } from '../../../../plugins/infra/server/lib/alerting/common/types';
+} from '../../../../plugins/infra/common/alerting/metrics';
+import { InfraSource } from '../../../../plugins/infra/common/source_configuration/source_configuration';
+import {
+  EvaluatedRuleParams,
+  evaluateRule,
+} from '../../../../plugins/infra/server/lib/alerting/metric_threshold/lib/evaluate_rule';
+import { FtrProviderContext } from '../../ftr_provider_context';
 import { DATES } from './constants';
 
 const { gauge, rate } = DATES['alert-test-data'];
